requirements.txt
openai==1.61.1
ping3==4.0.8
python-dotenv==1.0.1
streamlit==1.42.2